feat(core,cli): append to system prompt via CLI flag and config
- Add `--append-system-prompt <TEXT>` to `tui` and `exec`
- Add `append_system_prompt_file` (root/profile) to config
- Wire through `ConfigOverrides` → `Config` → `Session` → `Prompt`
- Append after base (`prompt.md`) and model-specific sections (e.g., apply_patch
)
- Ignore empty/whitespace appended text; CLI flag overrides file content
- Update docs (`config.md`) with replace vs append and precedence
- Add unit test for `get_full_instructions_with_append`
- Adjust `justfile` `fmt` to use `cargo fmt --all` (stable)
- Update MCP server constructors for new override field

Precedence:
1) CLI `--append-system-prompt` > 2) `append_system_prompt_file` (profile > root
) > 3) base (`prompt.md` or `experimental_instructions_file`)

Notes:
- No changes to `core/prompt.md`
- No changes related to sandbox env vars

+++ codex-rs/tui/src/bottom_pane/chat_composer.rs
@@ -1241,7 +1241,7 @@ mod tests {
                 composer.handle_paste(paste.clone());
                 composer
                     .textarea
-                    .set_cursor((placeholder.len() - pos_from_end) as usize);
+                    .set_cursor(placeholder.len() - pos_from_end);
                 composer.handle_key_event(KeyEvent::new(KeyCode::Backspace, KeyModifiers::NONE));
                 let result = (
                     composer.textarea.text().contains(&placeholder),
